% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cos_interp.r
\name{approx.cos}
\alias{approx.cos}
\title{Piecewise cosine interpolation}
\usage{
approx.cos(x, y = NULL, xout, n = 50, yleft, yright, rule = 1,
  ties = mean)
}
\arguments{
\item{x,y}{numeric vectors giving the coordinates of the points to be
interpolated.}

\item{xout}{an optional set of numeric values specifying where interpolation
is to take place.}

\item{n}{If xout is not specified, interpolation takes place at n equally
spaced points spanning the interval [\code{min(x)}, \code{max(x)}].}

\item{yleft}{the value to be returned when input \code{x} values are less
than \code{min(x)}. The default is defined by the value of rule given
below.}

\item{yright}{the value to be returned when input \code{x} values are greater
than \code{max(x)}. The default is defined by the value of rule given
below.}

\item{rule}{an integer (of length 1 or 2) describing how interpolation is to
take place outside the interval [\code{min(x)}, \code{max(x)}]. If rule is
\code{1} then \code{NA}s are returned for such points and if it is
\code{2}, the value at the closest data extreme is used. Use, e.g.,
\code{rule = 2:1}, if the left and right side extrapolation should differ.}

\item{ties}{Handling of tied \code{x} values. Either a function with a single
  vector argument returning a single number result or the string
  \code{"ordered"}.}
}
\description{
Returns a list of points which interpolate given data points. Similar to
\link[stats]{approx}.
}
\details{
This function is based on and functions similarly to
\link[stats]{approx} and \link[stats]{spline} .
}
\examples{
y <- rnorm(10)
x <- 1:10
plot(x, y)
lines(approx.cos(x, y, n=200))
}

