% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/decorations.r
\name{assign.scale.colors}
\alias{assign.scale.colors}
\title{Assign colors to value according to a scale}
\usage{
assign.scale.colors(x, scale.colors, scale.range)
}
\arguments{
\item{x}{values to be plotted}

\item{scale.colors}{a vector of colors}

\item{scale.range}{range for scale (vector with 2 numeric elements)}
}
\description{
\code{draw.scale} takes a vector of values, a color vector, and a range
vector and returns colors for plotting those values
}
\seealso{
\code{\link{draw.scale}} for drawing a color scale
}

