% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorations.r
\name{bin.ints}
\alias{bin.ints}
\title{Integer data binning}
\usage{
bin.ints(x, num.bins = 10)
}
\arguments{
\item{x}{data to be binned}

\item{num.bins}{number of bins to aim for}
}
\value{
\code{bin.ints} returns an ordered factor vector with appropriately
labeled levels.
}
\description{
Sometimes with integer data which is not uniformly distributed, a linear
color scale is not ideal. The function \code{bin.ints} converts an integer
vector into a factor vector where each level corresponds to a range of
integers and the elements of x are evenly distributed (as much as possible)
across the factor levels.
}
\details{
This is a convenience function that makesuse of \code{find.bins} and
\code{label.bins}.
}
\examples{
x <- rpois(100, 10)
x.bin <- bin.ints(x)


}
\seealso{
\code{\link{find.bins}} for determining bin breakpoints

\code{\link{label.bins}} for bin labels suitable for a legend
}

