% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/decorations.r
\name{draw.chrom.axis}
\alias{draw.chrom.axis}
\title{Draw a horizontal chromosome axis}
\usage{
draw.chrom.axis(start.pos, end.pos, chrom = NULL, label.chrom = TRUE,
  label.scale = TRUE, tick.length = 0.1, ...)
}
\arguments{
\item{start.pos}{Starting position on the chromosome (in bp)}

\item{end.pos}{Ending position on the chromosome (in bp)}

\item{...}{(optional) addition options to pass to \code{text} for drawing
labels}
}
\description{
\code{draw.chrom.axis} draws a chromosome axis with appropriate scale (Mb, kb, or bp).
}
\seealso{
\code{\link{draw.scale}} for drawing a color scale bar
}

