% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qq_plot.r
\name{fastqq}
\alias{fastqq}
\title{Fast QQ plots}
\usage{
fastqq(pvals, k = 2000, ...)
}
\description{
QQ plots with lots of points can be slow. Since most of the points overlap
we don't actually have to plot all of them (and in many cases the output
will be identical.)
}
\details{
This function will take a value k (default 2000) and sample that many
pvalues, then plot the log observed vs expected pvals in a nice way.

Additional arguments are passed to plot().
}

