% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorations.r
\name{find.bins}
\alias{find.bins}
\title{Find breakpoints for binning numeric values}
\usage{
find.bins(x, num.bins = 10)
}
\arguments{
\item{x}{data to be binned}

\item{num.bins}{number of bins to aim for}
}
\description{
Find breakpoints for binning numeric values
}
\examples{
x <- rpois(100, 10)
bins <- find.bins(x)
x.bin <- findInterval(x, bins)


}
\seealso{
\code{\link{label.bins}} for bin labels suitable for a legend

\code{\link{bin.ints}} for generating a factor with labels
}

