% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorations.r
\name{label.bins}
\alias{label.bins}
\title{Label bins generated by \code{find.bins()}}
\usage{
label.bins(x, bins, greedy = TRUE)
}
\arguments{
\item{x}{data to be binned}

\item{greedy}{(default TRUE) whether to include in labels values of x which
may not appear in the data. e.g. instead of \code{0, 1, 2-3, 5-6} we return
\code{0, 1, 2-4, 5-6}}

\item{bin}{binned data}
}
\description{
Label bins generated by \code{find.bins()}
}
\examples{
x <- rpois(100, 10)
bins <- find.bins(x)
x.bin <- findInterval(x, bins)
bin.labels <- label.bins(x, x.bin)

}
\seealso{
\code{\link{find.bins}} for determining bin breakpoints

\code{\link{bin.ints}} for generating a factor with labels
}

