% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biomart_gene_plot.r
\name{load.gene.mart}
\alias{load.gene.mart}
\title{Load gene database alternate version or organism.}
\usage{
load.gene.mart(version = "hg38", biomart.host, biomart.fields,
  biomart.filters, biomart.mart, biomart.dataset)
}
\description{
By default we will look up human (hg38) gene models. This function should be
called once after the package is loaded to use a different biomart database.
}
\examples{
# To use hg37 genomic coordinates
load.gene.mart('hg37')

# To use drosophila gene models
load.gene.mart('BDGP5')

# To use mouse gene models:
load.gene.mart(biomart.dataset='mmusculus_gene_ensembl',
               biomart.filters=NULL,
               biomart.filter.values=NULL)
}

