% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin_labels.r
\name{marginlabels}
\alias{marginlabels}
\title{Label points from the margin}
\usage{
marginlabels(x, y = NULL, labels = seq_along(x), margin = 4,
  col = "black", lty = 1, lwd = 1, pch = 1, pch.cex = 1, las = 2,
  rad = 0.15, ...)
}
\description{
Labels are sorted according to the axis along which they are labeled.
This could be made a lot better since this often ends up in crossed lines..
}
\examples{
y <- rnorm(100)
x <- runif(100)
plot(x, y, pch=20, bty='n')
label.pts <- tail(order(y), 10)
marginlabels(x[label.pts], y[label.pts], margin=3, lty=3, rad=0.05)
}

