% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/decorations.r
\name{prettybp}
\alias{prettybp}
\title{Print a bp length nicely}
\usage{
prettybp(n, round.digits = 2, signif.digits = NULL, space = TRUE)
}
\arguments{
\item{n}{base pairs}

\item{round.digits}{(optional) number of digits to round to. Default is 2. Set
to NULL to disable rounding. This value is ignored if signif.digits is not \code{NULL}.}

\item{signif.digits}{(optional) set this if rounding to significant digits
is preferred.}

\item{space}{Set to false to eliminate the space between the number and unit.}
}
\description{
\code{prettybp} returns a string representation of a base pair length with optional rounding
}
\seealso{
\code{\link{draw.chrom.axis}} for drawing a bp scaled x-axis
}

