% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorations.r
\name{splitter}
\alias{splitter}
\title{Calculate offsets for plotting y ~ x where y is continuous and x is categorical}
\usage{
splitter(y, x = NULL, rad = 0.025, scale = TRUE)
}
\description{
Calculate offsets for plotting y ~ x where y is continuous and x is categorical
}
\examples{
rx <- ceiling(runif(250, 0, 5))
ry <- rnorm(250, 0, 100)+runif(5, 0, 2000)[rx]
offset <- splitter(ry, rx)
plot(offset+as.numeric(rx), ry, pch=20, col=rainbow(5)[rx])
}

