% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_plots.r
\name{stackplot}
\alias{stackplot}
\title{Stacked plots}
\usage{
stackplot(y, x, spacing = 0.5, style = "normal", las = 1,
  y.scale = TRUE, lwd, lty, col, bty = "n", xaxt = par("xaxt"),
  yaxt = "n", yaxs = "i", xlab = "", ylab = "")
}
\arguments{
\item{y}{a matrix of values. Each row of the matrix will be one line in the
plot.}

\item{x}{(Optional) x axis values - length should correspond to columns in y matrix}

\item{spacing}{Spacing between lines}

\item{style}{normal or zero}
}
\description{
Stacked plots allow plotting multiple vectors of data without overlap in
minimal space.
}

