% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_plots.r
\name{y.scale.bar}
\alias{y.scale.bar}
\title{Y axis scale bar for stacked plots}
\usage{
y.scale.bar(length, pos = 0.5, abs.pos, las = par("las"), ...)
}
\arguments{
\item{length}{(Optional) if not supplied, the length will be the distance
between two ticks if the y axis had been plotted in the standard way (uses
\code{par('yaxp')}).}

\item{pos}{Vertical location for scale bar where 0 is at the bottom, 1 is at
the top, and 0.5 (default) is in the middle.}

\item{las}{Controls orientation of axis label text (See \code{par('las')})}

\item{...}{Other arguments to pass to axis (See \code{par()})}
}
\description{
In stacked plots the standard y axis doesn't really work - it is vertical
lengths of line segments that correspond to data values, not the absolute y
positions. This function draws a scale bar in the space where the yaxis
would normally be.
}
\seealso{
\code{\link{stack.plot}} for drawing stacked plots
}

