\name{zoom.plot-package}
\alias{zoom.plot-package}
\alias{zoom.plot}
\docType{package}
\title{
Various functions for plotting genomic data
}
\description{
Various functions for plotting genomic data
}
\details{
\tabular{ll}{
Package: \tab zoom.plot\cr
Type: \tab Package\cr
Version: \tab 0.1.3\cr
Date: \tab 2014-01-30\cr
License: \tab MIT\cr
}

}
\author{
Sushila Shenoy <sushila.alyssa.shenoy_gmail.com>
}
\keyword{ package }

