% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zoom_plot.r
\name{zoom.plot}
\alias{zoom.plot}
\title{Zoom in manhattan plot with gene models and LD plotted below.}
\usage{
zoom.plot(pvals, gene.id, all.snps, snp.id = NULL, genotypes,
  window = 1e+06, min.snps = 10, color = "maf", img.height = 700,
  img.width = 1200, img.prefix = "", save.image = TRUE)
}
\description{
Zoom in manhattan plot with gene models and LD plotted below.
}

