% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/decorations.r
\name{draw.scale}
\alias{draw.scale}
\title{Draw a color scale bar}
\usage{
draw.scale(scale.colors, scale.range, num.labs = 6, pos = "topleft",
  adj = NULL, horiz = TRUE, outside = FALSE, size = 2, ratio = 12,
  tick.length = 0.25, scale.offset = 0.5, label.offset = 0.1,
  box = TRUE, x.shift = 0, y.shift = 0)
}
\arguments{
\item{scale.colors}{a vector of colors}

\item{scale.range}{range for scale (vector with 2 numeric elements)}

\item{num.labs}{(optional) number of labels to draw}

\item{pos}{position for scale, either a keyword such as "top", "topleft" or a numeric vector length 2}

\item{adj}{controls the anchoring of the scale in respect to \code{pos}}

\item{horiz}{TRUE for horizontal (default) or FALSE for vertical bar}

\item{outside}{TRUE to display legend within the plotting area (default) or FALSE to place it outside}

\item{size}{approximate length in inches}

\item{ratio}{ratio of scale bar width to height}

\item{tick.length}{number between 0 and 1, controls how long the ticks are vs the color boxes.}

\item{scale.offset}{inset (outset) amount for legend that is inside (outside)}

\item{label.offset}{spacing between ticks and text labels}

\item{box}{this controls whether a black border is drawn around the colors or not.}

\item{x.shift}{(optional) adjust x position in units of scale width.}

\item{y.shift}{(optional) adjust y position in units of scale height.}
}
\description{
Adds a color scale legend to the current plot
with specified colors and range
}
\examples{
plot(1:5, 1:5, col=gray(0:4/5), pch=15)
draw.scale(gray(0:4/5), c(0, 1), pos='top')
draw.scale(gray(0:4/5), c(0, 1), pos='right', horiz=FALSE, outside=TRUE)
}
\seealso{
\code{\link{draw.chrom.axis}} for drawing x-axis
}

