% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/margin_labels.r
\name{marginlabels}
\alias{marginlabels}
\title{Label points from the margin}
\usage{
marginlabels(x, y = NULL, labels = seq_along(x), margin = 4,
  col = "black", lty = 1, lwd = 1, pch = 1, pch.cex = 1, las = 2,
  rad = 0.15, ...)
}
\description{
Label points from the margin
}

