% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qq_plot.r
\name{thin}
\alias{thin}
\title{Sampling for qq plots}
\usage{
thin(n, k = 2000)
}
\description{
Returns k indices between 1:n such that sampling is very dense at the start and much less dense at the end.
}

